requirements.txt
numpy~=1.24.3
pandas~=2.0.3
matplotlib~=3.7.2
scikit-learn~=1.3.0
pillow~=9.4.0
tensorflow==2.14.0
prettyprint==0.1.5
deeplake>=3.8.3
PySimpleGUI==4.60.5